chore: add unit tests for 'KinkedIR' model

+++ contracts/market/src/constants.rs
@@ -64,13 +64,13 @@ pub const DEFAULT_SLOPE1: i128 = 1;
 pub const DEFAULT_SLOPE2: i128 = 10;
 
 // ---- Interest Rate and Accrual 2 ----
-pub const DEFAULT_BASE_APR_BPS: u64 = 1; // 0.01%
-pub const DEFAULT_KINK1_UTILIZATION_RATIO_BPS: u64 = 7_000; // 70%
-pub const DEFAULT_KINK2_UTILIZATION_RATIO_BPS: u64 = 8_000; // 80%
+pub const DEFAULT_BASE_APR_BPS: i128 = 1; // 0.01%
+pub const DEFAULT_KINK1_UTILIZATION_RATIO_BPS: i128 = 7_000; // 70%
+pub const DEFAULT_KINK2_UTILIZATION_RATIO_BPS: i128 = 8_000; // 80%
 
-pub const DEFAULT_KINK1_APR_BPS: u64 = 3_000; // 30%
-pub const DEFAULT_KINK2_APR_BPS: u64 = 6_000; // 60%
-pub const DEFAULT_MAX_APR_BPS: u64 = 40_000; // 400%
+pub const DEFAULT_KINK1_APR_BPS: i128 = 3_000; // 30%
+pub const DEFAULT_KINK2_APR_BPS: i128 = 6_000; // 60%
+pub const DEFAULT_MAX_APR_BPS: i128 = 40_000; // 400%
 
 // ---- Deposit ----
